chore: move footer component
