+++ apps/web/src/app/layout.tsx
@@ -1,9 +1,9 @@
 import { DM_Sans } from "next/font/google";
 
 import "@repo/shadcn/styles/globals.css";
-import Footer from "@/components/Footer";
 import { Navbar } from "@/components/Navbar";
 import { Providers } from "@/components/Providers";
+import Footer from "@/components/sections/Footer";
 
 const fontSans = DM_Sans({
 	subsets: ["latin"],
